Extract mana coverage logic so it can be unit tested
manaCoverage.test.ts failed at import: pulling in ManaSourceSelectionUI drags in
the Zustand store, which reads localStorage while initialising the gameplay slice,
and that isn't defined under the node test environment.

Move the pure pieces — computeCoverage, pipColorOptions, pipGenericAmount, toPip,
isCovered, PipCoverage — into manaCoverage.ts, which imports only types. The
component keeps rendering; the test now imports the pure module and touches no
store.

No behaviour change; the hybrid-pip fix from the previous commit is unchanged.
